Refactor auttorisation as folder
